8.VideoGenHF/requirements.txt
# Core Dependencies for RTX 3050 6GB
torch>=2.0.0
torchvision>=0.15.0

# Diffusers and transformers
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0

# Streamlit frontend
streamlit>=1.28.0

# Image/Video processing
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
numpy>=1.24.0

# Memory optimization
safetensors>=0.4.0
